Fix independent critic value dtype under bf16 autocast
The softmax critic returns float32 (autocast forces softmax to fp32), but the
independent critic's torch.sigmoid stayed bfloat16 under the baseline's bf16
autocast, causing a float32-vs-bf16 index_put mismatch in the PPO value update
(current_values[idx] = new_values). Compute the independent value in float32 to
match the softmax critic's dtype contract. Adds a regression test. Found by the
GPU smoke (CPU smoke ran float32, so it didn't trigger autocast).

+++ python/owl/model/stateless_transformer_v1.py
@@ -1481,7 +1481,10 @@ class StatelessTransformerV1(BaseModelAPI):
             # no cross-player normalization, so non-zero-sum returns (e.g. the
             # ship-ratio reward) are representable. The second tuple element is
             # not a probability distribution, but matches the softmax tuple shape.
-            values = torch.sigmoid(logits).masked_fill(~still_playing, 0.0)
+            # Compute in float32 so the value dtype matches the softmax critic
+            # (softmax is autocast-forced to float32); downstream value buffers
+            # and index-puts assume float32 even under a bfloat16 autocast.
+            values = torch.sigmoid(logits.float()).masked_fill(~still_playing, 0.0)
             return values, values
         probabilities = masked_softmax(logits, still_playing, dim=-1)
         values = 2.0 * probabilities - 1.0
